+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## Release 4.26.1
 * Add failOnErrorMessages and failOnErrorCodes which are arrays of strings/regexes and integers, respectively. That prevent the retry logic from running. Effectively allowing for an early exit of specific failures.
+* Add option to provision BIG-IP modules using a transaction
 
 ## Release 4.26.0
 * Add BigIp.setHost method